fix: navigation improvement, tanstack examples separation
