+++ src/components/Tanstack/editable-cells/EditableCell.scss
@@ -1,7 +1,7 @@
 input {
   padding-left: var(--tds-spacing-element-16);
   height: 48px;
-  border-radius: none;
+  border-radius: 0;
   background-color: transparent;
   border: 0;
   border-bottom: 2px solid transparent;
